backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pymupdf
python-docx
sentence-transformers
faiss-cpu
python-dotenv
